+++ tests/test_users.py
@@ -49,7 +49,7 @@ class UserTests(TestCase):
                     }
         response = self.client().post('/api/v1/auth/signup', data=json.dumps(self.user))
         self.assertEqual(response.status_code, 400)
-        self.assertIn("Please provide a email address", str(response.data))
+        self.assertIn("Please provide an email address", str(response.data))
 
     # Test api can not create a user without an invalid email address
     def test_cannot_create_user_with_an_invalid_email(self):
